# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amazon Experts Backend API",
    "description": "Swagger/OpenAPI documentation for the Amazon Experts backend service. Platform APIs do not require x-tenant-id; tenant APIs require x-tenant-id in the request header.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/knowledge-bases/{knowledge_base_id}/docs/upload-url": {
      "post": {
        "tags": [
          "docs"
        ],
        "summary": "Create a document upload URL",
        "operationId": "create_upload_url_api_v1_knowledge_bases__knowledge_base_id__docs_upload_url_post",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadUrlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge-bases/{knowledge_base_id}/docs/upload-urls": {
      "post": {
        "tags": [
          "docs"
        ],
        "summary": "Create document upload URLs",
        "operationId": "create_upload_urls_api_v1_knowledge_bases__knowledge_base_id__docs_upload_urls_post",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadUrlsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadUrlsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge-bases/{knowledge_base_id}/docs/complete-upload": {
      "post": {
        "tags": [
          "docs"
        ],
        "summary": "Complete a document upload",
        "operationId": "complete_upload_api_v1_knowledge_bases__knowledge_base_id__docs_complete_upload_post",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge-bases/{knowledge_base_id}/docs/complete-uploads": {
      "post": {
        "tags": [
          "docs"
        ],
        "summary": "Complete document uploads",
        "operationId": "complete_uploads_api_v1_knowledge_bases__knowledge_base_id__docs_complete_uploads_post",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteUploadsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge-bases/{knowledge_base_id}/docs": {
      "get": {
        "tags": [
          "docs"
        ],
        "summary": "List knowledge base documents",
        "operationId": "list_documents_api_v1_knowledge_bases__knowledge_base_id__docs_get",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge-bases/{knowledge_base_id}/docs/{document_id}": {
      "get": {
        "tags": [
          "docs"
        ],
        "summary": "Get document details",
        "operationId": "get_document_api_v1_knowledge_bases__knowledge_base_id__docs__document_id__get",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "docs"
        ],
        "summary": "Update document metadata",
        "operationId": "update_document_api_v1_knowledge_bases__knowledge_base_id__docs__document_id__patch",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "docs"
        ],
        "summary": "Delete a document",
        "operationId": "delete_document_api_v1_knowledge_bases__knowledge_base_id__docs__document_id__delete",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge-bases/{knowledge_base_id}/docs/{document_id}/download-url": {
      "get": {
        "tags": [
          "docs"
        ],
        "summary": "Get a document download URL",
        "operationId": "get_download_url_api_v1_knowledge_bases__knowledge_base_id__docs__document_id__download_url_get",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Knowledge Base Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchItemError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "BatchItemError"
      },
...
      "CompleteUploadRequest": {
        "properties": {
          "uploadSessionId": {
            "type": "string",
            "title": "Uploadsessionid"
          },
          "etag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Etag"
          },
          "fileSizeBytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filesizebytes"
          }
        },
        "type": "object",
        "required": [
          "uploadSessionId"
        ],
        "title": "CompleteUploadRequest"
      },
      "CompleteUploadResult": {
        "properties": {
          "uploadSessionId": {
            "type": "string",
            "title": "Uploadsessionid"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "document": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Document"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BatchItemError"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "uploadSessionId",
          "status"
        ],
        "title": "CompleteUploadResult"
      },
      "CompleteUploadsRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CompleteUploadRequest"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CompleteUploadsRequest"
      },
      "CompleteUploadsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CompleteUploadResult"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CompleteUploadsResponse"
      },
...
      "Document": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "knowledgeBaseId": {
            "type": "string",
            "title": "Knowledgebaseid"
          },
          "fileName": {
            "type": "string",
            "title": "Filename"
          },
          "fileType": {
            "type": "string",
            "enum": [
              "pdf",
              "docx",
              "pptx",
              "xlsx",
              "md",
              "txt",
              "html",
              "csv",
              "json"
            ],
            "title": "Filetype"
          },
          "mimeType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mimetype"
          },
          "fileSizeBytes": {
            "type": "integer",
            "title": "Filesizebytes"
          },
          "storageKey": {
            "type": "string",
            "title": "Storagekey"
          },
          "contentHash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contenthash"
          },
          "parseStatus": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "ready",
              "failed"
            ],
            "title": "Parsestatus"
          },
          "indexStatus": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "ready",
              "failed",
              "stale"
            ],
            "title": "Indexstatus"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "knowledgeBaseId",
          "fileName",
          "fileType",
          "fileSizeBytes",
          "storageKey",
          "parseStatus",
          "indexStatus",
          "createdAt",
          "updatedAt"
        ],
        "title": "Document"
      },
      "DocumentListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Document"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "DocumentListResponse"
      },
      "DownloadUrlResponse": {
        "properties": {
          "method": {
            "type": "string",
            "const": "GET",
            "title": "Method",
            "default": "GET"
          },
          "downloadUrl": {
            "type": "string",
            "title": "Downloadurl"
          },
          "expiresAt": {
            "type": "string",
            "title": "Expiresat"
          }
        },
        "type": "object",
        "required": [
          "downloadUrl",
          "expiresAt"
        ],
        "title": "DownloadUrlResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateDocumentRequest": {
        "properties": {
          "fileName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "title": "UpdateDocumentRequest"
      },
...
      "UploadUrlRequest": {
        "properties": {
          "fileName": {
            "type": "string",
            "title": "Filename"
          },
          "mimeType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mimetype"
          },
          "fileSizeBytes": {
            "type": "integer",
            "title": "Filesizebytes"
          },
          "contentHash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contenthash"
          }
        },
        "type": "object",
        "required": [
          "fileName",
          "fileSizeBytes"
        ],
        "title": "UploadUrlRequest"
      },
      "UploadUrlResponse": {
        "properties": {
          "uploadSessionId": {
            "type": "string",
            "title": "Uploadsessionid"
          },
          "documentId": {
            "type": "string",
            "title": "Documentid"
          },
          "method": {
            "type": "string",
            "const": "PUT",
            "title": "Method",
            "default": "PUT"
          },
          "uploadUrl": {
            "type": "string",
            "title": "Uploadurl"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Headers"
          },
          "objectKey": {
            "type": "string",
            "title": "Objectkey"
          },
          "expiresAt": {
            "type": "string",
            "title": "Expiresat"
          }
        },
        "type": "object",
        "required": [
          "uploadSessionId",
          "documentId",
          "uploadUrl",
          "objectKey",
          "expiresAt"
        ],
        "title": "UploadUrlResponse"
      },
      "UploadUrlResult": {
        "properties": {
          "fileName": {
            "type": "string",
            "title": "Filename"
          },
          "status": {
            "type": "string",
            "enum": [
              "created",
              "failed"
            ],
            "title": "Status"
          },
          "upload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UploadUrlResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BatchItemError"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "fileName",
          "status"
        ],
        "title": "UploadUrlResult"
      },
      "UploadUrlsRequest": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/UploadUrlRequest"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "UploadUrlsRequest"
      },
      "UploadUrlsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UploadUrlResult"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "UploadUrlsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}